requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
anthropic==0.39.0
openai==1.35.0
watchdog==4.0.2
gitpython==3.1.42
colorama==0.4.6
pydantic==2.5.3
aiofiles==23.2.1
mss==10.1.0
pypdf==5.1.0
